problem 4 done

+++ src/plugins/pwscf.py
@@ -24,7 +24,7 @@ def qe_value_map(value):
             return '.true.'
         else:
             return '.false.'
-    elif isinstance(value, (float, numpy.float)) or isinstance(value, (int, numpy.int)):
+    elif isinstance(value, (float, float)) or isinstance(value, (int, int)):
         return str(value)
     elif isinstance(value, str):
         return "'{}'".format(value)
